LPD-92999 Single-source the Help & Support specification keys
Deriving the tab visibility keys from SUPPORT_LINKS keeps the visibility check and the rendered rows from drifting apart when a support field is added or removed.
